refactor to avoid cyclic imports
