+++ monoloco/eval/generate_kitti.py
@@ -14,9 +14,10 @@ import torch
 from ..network import Loco
 from ..network.process import preprocess_pifpaf
 from ..network.geom_baseline import geometric_coordinates
-from ..utils import get_keypoints, pixel_to_camera, factory_file, factory_basename, make_new_directory, get_category, \
+from ..utils import get_keypoints, pixel_to_camera, factory_basename, make_new_directory, get_category, \
     xyz_from_distance, read_and_rewrite
 from .stereo_baselines import baselines_association
+from ..prep import factory_file
 from .reid_baseline import get_reid_features, ReID
