+++ monoloco/run.py
@@ -131,7 +131,7 @@ def main():
             prep = PreprocessNuscenes(args.dir_ann, args.dir_nuscenes, args.dataset, args.iou_min)
             prep.run()
         else:
-            from .prep.prep_kitti import PreprocessKitti
+            from .prep.preprocess_kitti import PreprocessKitti
             prep = PreprocessKitti(args.dir_ann, mode=args.mode, iou_min=args.iou_min)
             if args.activity:
                 prep.prep_activity()
