+++ monoloco/utils/__init__.py
@@ -1,12 +1,13 @@
 
-from .iou import get_iou_matches, reorder_matches, get_iou_matrix, get_iou_matches_matrix
-from .misc import get_task_error, get_pixel_error, append_cluster, open_annotations, make_new_directory,\
+from .iou import get_iou_matches, reorder_matches, get_iou_matrix, get_iou_matches_matrix, get_category, \
+    open_annotations
+from .misc import get_task_error, get_pixel_error, append_cluster,  make_new_directory,\
     normalize_hwl, average
-from .kitti import check_conditions, get_difficulty, split_training, parse_ground_truth, get_calibration, \
-    factory_basename, factory_file, get_category, read_and_rewrite, find_cluster
+from .kitti import check_conditions, get_difficulty, split_training, get_calibration, \
+    factory_basename, read_and_rewrite, find_cluster
 from .camera import xyz_from_distance, get_keypoints, pixel_to_camera, project_3d, open_image, correct_angle,\
     to_spherical, to_cartesian, back_correct_angles, project_to_pixels
 from .logs import set_logger
-from ..utils.nuscenes import select_categories
-from ..utils.stereo import mask_joint_disparity, average_locations, extract_stereo_matches, \
+from .nuscenes import select_categories
+from .stereo import mask_joint_disparity, average_locations, extract_stereo_matches, \
     verify_stereo, disparity_to_depth
