+++ monoloco/utils/kitti.py
@@ -1,5 +1,4 @@
 
-import math
 import os
 import glob
 
@@ -129,50 +128,6 @@ def split_training(names_gt, path_train, path_val):
     return set_train, set_val
 
 
-def parse_ground_truth(path_gt, category, spherical=False, verbose=False):
-    """Parse KITTI ground truth files"""
-    from ..utils import correct_angle, to_spherical
-
-    boxes_gt = []
-    ys = []
-    truncs_gt = []  # Float from 0 to 1
-    occs_gt = []  # Either 0,1,2,3 fully visible, partly occluded, largely occluded, unknown
-    lines = []
-
-    with open(path_gt, "r") as f_gt:
-        for line_gt in f_gt:
-            line = line_gt.split()
-            if check_conditions(line_gt, category, method='gt'):
-                truncs_gt.append(float(line[1]))
-                occs_gt.append(int(line[2]))
-                boxes_gt.append([float(x) for x in line[4:8]])
-                xyz = [float(x) for x in line[11:14]]
-                hwl = [float(x) for x in line[8:11]]
-                dd = float(math.sqrt(xyz[0] ** 2 + xyz[1] ** 2 + xyz[2] ** 2))
-                yaw = float(line[14])
-                assert - math.pi <= yaw <= math.pi
-                alpha = float(line[3])
-                sin, cos, yaw_corr = correct_angle(yaw, xyz)
-                assert min(abs(-yaw_corr - alpha), (abs(yaw_corr - alpha))) < 0.15, "more than 10 degrees of error"
-                if spherical:
-                    rtp = to_spherical(xyz)
-                    loc = rtp[1:3] + xyz[2:3] + rtp[0:1]  # [theta, psi, z, r]
-                else:
-                    loc = xyz + [dd]
-                # cat = 0 if line[0] in ('Pedestrian', 'Person_sitting') else 1
-                if line[0] in ('Pedestrian', 'Person_sitting'):
-                    cat = 0
-                else:
-                    cat = 1
-                output = loc + hwl + [sin, cos, yaw, cat]
-                ys.append(output)
-                if verbose:
-                    lines.append(line_gt)
-    if verbose:
-        return boxes_gt, ys, truncs_gt, occs_gt, lines
-    return boxes_gt, ys, truncs_gt, occs_gt
-
-
 def factory_basename(dir_ann, dir_gt):
     """ Return all the basenames in the annotations folder corresponding to validation images"""
 
@@ -189,64 +144,6 @@ def factory_basename(dir_ann, dir_gt):
     set_val = set_basename.intersection(set_val_gt)
     assert set_val, " Missing json annotations file to create txt files for KITTI datasets"
     return set_val
-
-
-def factory_file(path_calib, dir_ann, basename, mode='left'):
-    """Choose the annotation and the calibration files. Stereo option with ite = 1"""
-
-    assert mode in ('left', 'right')
-    p_left, p_right = get_calibration(path_calib)
-
-    if mode == 'left':
-        kk, tt = p_left[:]
-        path_ann = os.path.join(dir_ann, basename + '.png.predictions.json')
-
-    else:
-        kk, tt = p_right[:]
-        path_ann = os.path.join(dir_ann + '_right', basename + '.png.predictions.json')
-
-    from ..utils import open_annotations
-    annotations = open_annotations(path_ann)
-
-    return annotations, kk, tt
-
-
-def get_category(keypoints, path_byc):
-    """Find the category for each of the keypoints"""
-
-    from ..utils import open_annotations
-    dic_byc = open_annotations(path_byc)
-    boxes_byc = dic_byc['boxes'] if dic_byc else []
-    boxes_ped = make_lower_boxes(keypoints)
-
-    matches = get_matches_bikes(boxes_ped, boxes_byc)
-    list_byc = [match[0] for match in matches]
-    categories = [1.0 if idx in list_byc else 0.0 for idx, _ in enumerate(boxes_ped)]
-    return categories
-
-
-def get_matches_bikes(boxes_ped, boxes_byc):
-    from . import get_iou_matches_matrix
-    matches = get_iou_matches_matrix(boxes_ped, boxes_byc, thresh=0.15)
-    matches_b = []
-    for idx, idx_byc in matches:
-        box_ped = boxes_ped[idx]
-        box_byc = boxes_byc[idx_byc]
-        width_ped = box_ped[2] - box_ped[0]
-        width_byc = box_byc[2] - box_byc[0]
-        center_ped = (box_ped[2] + box_ped[0]) / 2
-        center_byc = (box_byc[2] + box_byc[0]) / 2
-        if abs(center_ped - center_byc) < min(width_ped, width_byc) / 4:
-            matches_b.append((idx, idx_byc))
-    return matches_b
-
-
-def make_lower_boxes(keypoints):
-    lower_boxes = []
-    keypoints = np.array(keypoints)
-    for kps in keypoints:
-        lower_boxes.append([min(kps[0, 9:]), min(kps[1, 9:]), max(kps[0, 9:]), max(kps[1, 9:])])
-    return lower_boxes
 
 
 def read_and_rewrite(path_orig, path_new):
