+++ tests/test_utils.py
@@ -6,7 +6,7 @@ sys.path.insert(0, os.path.join('..', 'monoloco'))
 
 
 def test_iou():
-    from stereoloco.utils import get_iou_matrix
+    from monoloco.utils import get_iou_matrix
     boxes_pred = [[1, 100, 1, 200]]
     boxes_gt = [[100., 120., 150., 160.],[12, 110, 130., 160.]]
     iou_matrix = get_iou_matrix(boxes_pred, boxes_gt)
@@ -14,7 +14,7 @@ def test_iou():
 
 
 def test_pixel_to_camera():
-    from stereoloco.utils import pixel_to_camera
+    from monoloco.utils import pixel_to_camera
     kk = [[718.3351, 0., 600.3891], [0., 718.3351, 181.5122], [0., 0., 1.]]
     zz = 10
     uv_vector = [1000., 400.]
